Uitgebreide versie van het spel
-Verschillende schermen toegevoegd (home, game en game over).
-Puntentelling toegevoegd.
