+++ .idea/gradle.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="GradleMigrationSettings" migrationVersion="1" />
+  <component name="GradleSettings">
+    <option name="linkedExternalProjectsSettings">
+      <GradleProjectSettings>
+        <option name="distributionType" value="DEFAULT_WRAPPED" />
+        <option name="externalProjectPath" value="$PROJECT_DIR$/libgdx-test" />
+        <option name="modules">
+          <set>
+            <option value="$PROJECT_DIR$/libgdx-test" />
+            <option value="$PROJECT_DIR$/libgdx-test/core" />
+            <option value="$PROJECT_DIR$/libgdx-test/desktop" />
+            <option value="$PROJECT_DIR$/libgdx-test/html" />
+            <option value="$PROJECT_DIR$/libgdx-test/ios" />
+          </set>
+        </option>
+      </GradleProjectSettings>
+    </option>
+  </component>
+</project>
