+++ libgdx-test/core/assets/skin/craftacular-ui.json
@@ -1,0 +1,160 @@
+{
+com.badlogic.gdx.graphics.g2d.BitmapFont: {
+	bold: {
+		file: font-bold-export.fnt
+	}
+	font: {
+		file: font-export.fnt
+	}
+	title: {
+		file: font-title-export.fnt
+	}
+}
+com.badlogic.gdx.graphics.Color: {
+	gray: {
+		r: 0.55
+		g: 0.55
+		b: 0.55
+		a: 1
+	}
+	green: {
+		r: 0.03529522
+		g: 0.77666664
+		b: 0
+		a: 1
+	}
+	white: {
+		r: 1
+		g: 1
+		b: 1
+		a: 1
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.Skin$TintedDrawable: {
+	gray: {
+		name: white
+		color: {
+			r: 0.73666674
+			g: 0.73666674
+			b: 0.73666674
+			a: 1
+		}
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.Button$ButtonStyle: {
+	default: {
+		up: button
+		over: button-hover
+		disabled: button-disabled
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.ImageButton$ImageButtonStyle: {
+	default: {
+		up: button
+		over: button-hover
+		disabled: button-disabled
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.ImageTextButton$ImageTextButtonStyle: {
+	default: {
+		font: font
+		up: button
+		over: button-hover
+		disabled: button-disabled
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.Label$LabelStyle: {
+	default: {
+		font: font
+	}
+	dim: {
+		font: font
+		fontColor: gray
+	}
+	bold: {
+		font: bold
+		background: underline
+	}
+	title: {
+		font: title
+	}
+	xp: {
+		font: bold
+		fontColor: green
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.ProgressBar$ProgressBarStyle: {
+	default-horizontal: {
+		background: slider
+		knobBefore: slider-knob
+	}
+	default-vertical: {
+		background: slider
+		knob: slider-knob
+	}
+	hunger: {
+		background: meat-bg
+		knobAfter: meat
+	}
+	health: {
+		background: heart-bg
+		knobBefore: heart
+	}
+	xp: {
+		background: xp-bg
+		knobBefore: xp
+	}
+	armor: {
+		background: armor-bg
+		knobBefore: armor
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.ScrollPane$ScrollPaneStyle: {
+	default: {
+		background: scroll-bg
+		hScroll: scroll-horizontal
+		hScrollKnob: scroll-knob-horizontal
+		vScroll: scroll-vertical
+		vScrollKnob: scroll-knob-vertical
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.Slider$SliderStyle: {
+	default-horizontal: {
+		background: slider
+		knob: slider-knob
+	}
+	default-vertical: {
+		background: slider
+		knob: slider-knob
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.TextButton$TextButtonStyle: {
+	default: {
+		font: font
+		up: button
+		over: button-hover
+		disabled: button-disabled
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.TextField$TextFieldStyle: {
+	default: {
+		font: font
+		fontColor: white
+		background: text-field
+		cursor: cursor
+		selection: gray
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.TextTooltip$TextTooltipStyle: {
+	default: {
+		label: default
+		background: tooltip
+	}
+}
+com.badlogic.gdx.scenes.scene2d.ui.Window$WindowStyle: {
+	default: {
+		background: window
+		titleFont: font
+	}
+}
+}
