+++ libgdx-test/core/src/com/badlogic/drop/Drop.java
@@ -1,202 +1,30 @@
 package com.badlogic.drop;
 
-import java.io.IOException;
-import java.net.DatagramPacket;
-import java.net.DatagramSocket;
-import java.net.SocketException;
-import java.net.SocketTimeoutException;
-import java.util.Iterator;
-import com.fazecast.jSerialComm.*;
-import arduino.*;
-import com.badlogic.gdx.ApplicationAdapter;
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Input.Keys;
-import com.badlogic.gdx.audio.Music;
-import com.badlogic.gdx.audio.Sound;
-import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.Game;
+import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.math.MathUtils;
-import com.badlogic.gdx.math.Rectangle;
-import com.badlogic.gdx.math.Vector3;
-import com.badlogic.gdx.utils.Array;
-import com.badlogic.gdx.utils.ScreenUtils;
-import com.badlogic.gdx.utils.TimeUtils;
 
-public class Drop extends ApplicationAdapter {
-    private Texture dropImage;
-    private Texture bucketImage;
-    private Sound dropSound;
-    private Music rainMusic;
-    private SpriteBatch batch;
-    private OrthographicCamera camera;
-    private Rectangle bucket;
-    private Array<Rectangle> raindrops;
-    private long lastDropTime;
-    private long diff, start = System.currentTimeMillis();
-    public String input = "";
-    final SerialPort comPort = SerialPort.getCommPort("COM3");
 
-    int MAX_LEN = 2048;
-    int localPortNum = 22;
-    DatagramSocket mySocket;
-    byte[] buffer;
-    DatagramPacket packet;
+public class Drop extends Game {
 
-    @Override
+    public SpriteBatch batch;
+    public BitmapFont font;
+    public MainMenuScreen screen;
+
     public void create() {
-        try {
-            mySocket = new DatagramSocket(localPortNum);
-        } catch (SocketException e) {
-            e.printStackTrace();
-        }
-        buffer = new byte[MAX_LEN];
-        packet = new DatagramPacket(buffer, MAX_LEN);
-        comPort.openPort();
-        comPort.addDataListener(new SerialPortDataListener() {
-            @Override
-            public int getListeningEvents() { return SerialPort.LISTENING_EVENT_DATA_AVAILABLE; }
-            @Override
-            public void serialEvent(SerialPortEvent event) {
-                if (event.getEventType() != SerialPort.LISTENING_EVENT_DATA_AVAILABLE)
-                    return; else {
-                }
-                byte[] newData = new byte[comPort.bytesAvailable()];
-                int numRead = comPort.readBytes(newData, newData.length);
-                String result = new String(newData);
-                input = result;
-            }
-        });
-
-        // load the images for the droplet and the bucket, 64x64 pixels each
-        dropImage = new Texture(Gdx.files.internal("drop.png"));
-        bucketImage = new Texture(Gdx.files.internal("bucket.png"));
-
-        // load the drop sound effect and the rain background "music"
-        dropSound = Gdx.audio.newSound(Gdx.files.internal("drop.wav"));
-        rainMusic = Gdx.audio.newMusic(Gdx.files.internal("rain.mp3"));
-
-        // start the playback of the background music immediately
-        rainMusic.setLooping(true);
-        rainMusic.play();
-
-        // create the camera and the SpriteBatch
-        camera = new OrthographicCamera();
-        camera.setToOrtho(false, 800, 480);
         batch = new SpriteBatch();
-
-        // create a Rectangle to logically represent the bucket
-        bucket = new Rectangle();
-        bucket.x = 800 / 2 - 64 / 2; // center the bucket horizontally
-        bucket.y = 20; // bottom left corner of the bucket is 20 pixels above the bottom screen edge
-        bucket.width = 64;
-        bucket.height = 64;
-
-        // create the raindrops array and spawn the first raindrop
-        raindrops = new Array<Rectangle>();
-        spawnRaindrop();
-
-
-
+        font = new BitmapFont(); // use libGDX's default Arial font
+        screen = new MainMenuScreen(this);
+        this.setScreen(screen);
     }
 
-    private void spawnRaindrop() {
-        Rectangle raindrop = new Rectangle();
-        raindrop.x = MathUtils.random(0, 800-64);
-        raindrop.y = 480;
-        raindrop.width = 64;
-        raindrop.height = 64;
-        raindrops.add(raindrop);
-        lastDropTime = TimeUtils.nanoTime();
-    }
-
-    @Override
     public void render() {
-        try {
-            mySocket.setSoTimeout(1);
-            mySocket.receive(packet);
-            String message = new String(buffer);
-            input = message;
-        } catch (IOException e) {
-        }
-        int fps = 60;
-        // clear the screen with a dark blue color. The
-        // arguments to clear are the red, green
-        // blue and alpha component in the range [0,1]
-        // of the color to be used to clear the screen.
-        ScreenUtils.clear(0, 0, 0.2f, 1);
-
-        // tell the camera to update its matrices.
-        camera.update();
-
-        // tell the SpriteBatch to render in the
-        // coordinate system specified by the camera.
-        batch.setProjectionMatrix(camera.combined);
-
-        // begin a new batch and draw the bucket and
-        // all drops
-        batch.begin();
-        batch.draw(bucketImage, bucket.x, bucket.y);
-        for(Rectangle raindrop: raindrops) {
-            batch.draw(dropImage, raindrop.x, raindrop.y);
-        }
-        batch.end();
-
-        // process user input
-        if(Gdx.input.isTouched()) {
-            Vector3 touchPos = new Vector3();
-            touchPos.set(Gdx.input.getX(), Gdx.input.getY(), 0);
-            camera.unproject(touchPos);
-            bucket.x = touchPos.x - 64 / 2;
-        }
-
-        if(input.contains("left") || input.contains("A")) bucket.x -= 400 * Gdx.graphics.getDeltaTime();
-        if(input.contains("right") || input.contains("D")) bucket.x += 400 * Gdx.graphics.getDeltaTime();
-        input = "";
-        // make sure the bucket stays within the screen bounds
-        if(bucket.x < 0) bucket.x = 0;
-        if(bucket.x > 800 - 64) bucket.x = 800 - 64;
-
-        // check if we need to create a new raindrop
-        if(TimeUtils.nanoTime() - lastDropTime > 1000000000) spawnRaindrop();
-
-        // move the raindrops, remove any that are beneath the bottom edge of
-        // the screen or that hit the bucket. In the latter case we play back
-        // a sound effect as well.
-        for (Iterator<Rectangle> iter = raindrops.iterator(); iter.hasNext(); ) {
-            Rectangle raindrop = iter.next();
-            raindrop.y -= 200 * Gdx.graphics.getDeltaTime();
-            if(raindrop.y + 64 < 0) iter.remove();
-            if(raindrop.overlaps(bucket)) {
-                dropSound.play();
-                iter.remove();
-            }
-        }
-        sleep(fps);
+        super.render(); // important!
     }
 
-    public void sleep(int fps) {
-        if(fps>0){
-            diff = System.currentTimeMillis() - start;
-            long targetDelay = 1000/fps;
-            if (diff < targetDelay) {
-                try{
-                    Thread.sleep(targetDelay - diff);
-                } catch (InterruptedException e) {}
-            }
-            start = System.currentTimeMillis();
-        }
-    }
-
-    @Override
     public void dispose() {
-        // dispose of all the native resources
-        dropImage.dispose();
-        bucketImage.dispose();
-        dropSound.dispose();
-        rainMusic.dispose();
         batch.dispose();
-        comPort.closePort();
-        mySocket.close();
+        font.dispose();
     }
+
 }
