+++ libgdx-test/core/src/com/badlogic/drop/GameOverScreen.java
@@ -1,0 +1,71 @@
+package com.badlogic.drop;
+
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.utils.ScreenUtils;
+
+public class GameOverScreen implements Screen {
+
+    final Drop game;
+    int points;
+
+    OrthographicCamera camera;
+
+    public GameOverScreen(final Drop game, int points) {
+        this.game = game;
+        this.points = points;
+
+        camera = new OrthographicCamera();
+        camera.setToOrtho(false, 800, 480);
+    }
+
+    @Override
+    public void show() {
+
+    }
+
+    public void render(float delta) {
+        ScreenUtils.clear(0, 0, 0.2f, 1);
+
+        camera.update();
+        game.batch.setProjectionMatrix(camera.combined);
+
+        game.batch.begin();
+        game.font.draw(game.batch, String.format("Game Over   You have %s points", points), 100, 150);
+        game.font.draw(game.batch, "Tap anywhere to restart!", 100, 100);
+        game.batch.end();
+
+        if (Gdx.input.isTouched()) {
+            game.setScreen(new GameScreen(game));
+            dispose();
+        }
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+    //...Rest of class omitted for succinctness.
+
+}
