+++ libgdx-test/core/src/com/badlogic/drop/GameScreen.java
@@ -1,0 +1,188 @@
+package com.badlogic.drop;
+
+import java.io.IOException;
+import java.net.DatagramPacket;
+import java.net.DatagramSocket;
+import java.net.SocketException;
+import java.net.SocketTimeoutException;
+import java.util.Iterator;
+
+import com.badlogic.gdx.Screen;
+import com.fazecast.jSerialComm.*;
+import arduino.*;
+import com.badlogic.gdx.ApplicationAdapter;
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input.Keys;
+import com.badlogic.gdx.audio.Music;
+import com.badlogic.gdx.audio.Sound;
+import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.badlogic.gdx.math.MathUtils;
+import com.badlogic.gdx.math.Rectangle;
+import com.badlogic.gdx.math.Vector3;
+import com.badlogic.gdx.utils.Array;
+import com.badlogic.gdx.utils.ScreenUtils;
+import com.badlogic.gdx.utils.TimeUtils;
+
+public class GameScreen implements Screen {
+    final Drop game;
+
+    Texture dropImage;
+    Texture bucketImage;
+    Sound dropSound;
+    Music rainMusic;
+    OrthographicCamera camera;
+    Rectangle bucket;
+    Array<Rectangle> raindrops;
+    long lastDropTime;
+    int dropsGathered;
+
+    public GameScreen(final Drop game) {
+        this.game = game;
+
+        // load the images for the droplet and the bucket, 64x64 pixels each
+        dropImage = new Texture(Gdx.files.internal("drop.png"));
+        bucketImage = new Texture(Gdx.files.internal("bucket.png"));
+
+        // load the drop sound effect and the rain background "music"
+        dropSound = Gdx.audio.newSound(Gdx.files.internal("drop.wav"));
+        rainMusic = Gdx.audio.newMusic(Gdx.files.internal("rain.mp3"));
+        rainMusic.setLooping(true);
+
+        // create the camera and the SpriteBatch
+        camera = new OrthographicCamera();
+        camera.setToOrtho(false, 800, 480);
+
+        // create a Rectangle to logically represent the bucket
+        bucket = new Rectangle();
+        bucket.x = 800 / 2 - 64 / 2; // center the bucket horizontally
+        bucket.y = 20; // bottom left corner of the bucket is 20 pixels above
+        // the bottom screen edge
+        bucket.width = 64;
+        bucket.height = 64;
+
+        // create the raindrops array and spawn the first raindrop
+        raindrops = new Array<Rectangle>();
+        spawnRaindrop();
+
+    }
+
+    private void spawnRaindrop() {
+        Rectangle raindrop = new Rectangle();
+        raindrop.x = MathUtils.random(0, 800 - 64);
+        raindrop.y = 480;
+        raindrop.width = 64;
+        raindrop.height = 64;
+        raindrops.add(raindrop);
+        lastDropTime = TimeUtils.nanoTime();
+    }
+
+    @Override
+    public void render(float delta) {
+        //Ophalen button data van Raspberry Pi
+        try {
+            game.screen.mySocket.setSoTimeout(1);
+            game.screen.mySocket.receive(game.screen.packet);
+            String message = new String(game.screen.buffer);
+            game.screen.input = message;
+        } catch (IOException ignored) {
+        }
+
+        // clear the screen with a dark blue color. The
+        // arguments to clear are the red, green
+        // blue and alpha component in the range [0,1]
+        // of the color to be used to clear the screen.
+        ScreenUtils.clear(0, 0, 0.2f, 1);
+
+        // tell the camera to update its matrices.
+        camera.update();
+
+        // tell the SpriteBatch to render in the
+        // coordinate system specified by the camera.
+        game.batch.setProjectionMatrix(camera.combined);
+
+        // begin a new batch and draw the bucket and
+        // all drops
+        game.batch.begin();
+        game.font.draw(game.batch, "Drops Collected: " + dropsGathered, 0, 480);
+        game.batch.draw(bucketImage, bucket.x, bucket.y, bucket.width, bucket.height);
+        for (Rectangle raindrop : raindrops) {
+            game.batch.draw(dropImage, raindrop.x, raindrop.y);
+        }
+        game.batch.end();
+
+        // process user input
+        if (Gdx.input.isTouched()) {
+            Vector3 touchPos = new Vector3();
+            touchPos.set(Gdx.input.getX(), Gdx.input.getY(), 0);
+            camera.unproject(touchPos);
+            bucket.x = touchPos.x - 64 / 2;
+        }
+        if(game.screen.input.contains("left") || game.screen.input.contains("A")) bucket.x -= 400 * Gdx.graphics.getDeltaTime();
+        if(game.screen.input.contains("right") || game.screen.input.contains("D")) bucket.x += 400 * Gdx.graphics.getDeltaTime();
+        game.screen.input = "";
+
+        // make sure the bucket stays within the screen bounds
+        if (bucket.x < 0)
+            bucket.x = 0;
+        if (bucket.x > 800 - 64)
+            bucket.x = 800 - 64;
+
+        // check if we need to create a new raindrop
+        if (TimeUtils.nanoTime() - lastDropTime > 1000000000)
+            spawnRaindrop();
+
+        // move the raindrops, remove any that are beneath the bottom edge of
+        // the screen or that hit the bucket. In the later case we increase the
+        // value our drops counter and add a sound effect.
+        Iterator<Rectangle> iter = raindrops.iterator();
+        while (iter.hasNext()) {
+            Rectangle raindrop = iter.next();
+            raindrop.y -= 200 * Gdx.graphics.getDeltaTime();
+            if (raindrop.y < 0) {
+                game.setScreen(new GameOverScreen(game, dropsGathered));
+                dispose();
+            }
+            if (raindrop.y + 64 < 0)
+                iter.remove();
+            if (raindrop.overlaps(bucket)) {
+                dropsGathered++;
+                dropSound.play();
+                iter.remove();
+            }
+        }
+    }
+
+    @Override
+    public void resize(int width, int height) {
+    }
+
+    @Override
+    public void show() {
+        // start the playback of the background music
+        // when the screen is shown
+        rainMusic.play();
+    }
+
+    @Override
+    public void hide() {
+    }
+
+    @Override
+    public void pause() {
+    }
+
+    @Override
+    public void resume() {
+    }
+
+    @Override
+    public void dispose() {
+        dropImage.dispose();
+        bucketImage.dispose();
+        dropSound.dispose();
+        rainMusic.dispose();
+    }
+
+}
